AP-15434: Make sizeof return long
AP-15434 (Adjust ChunkSize by analyzing memory-footprint)

(This commit was moved from knime-core-arrow. Original id: 462aa70984d46d97ae510586e7be01b81e8baf38)

+++ org.knime.core.columnar.arrow.tests/src/test/java/org/knime/core/columnar/arrow/ArrowTestUtils.java
@@ -178,7 +178,7 @@ public final class ArrowTestUtils {
         }
 
         @Override
-        public int sizeOf() {
+        public long sizeOf() {
             // Does not matter for this test data
             return 100;
         }
@@ -271,7 +271,7 @@ public final class ArrowTestUtils {
         }
 
         @Override
-        public int sizeOf() {
+        public long sizeOf() {
             // Does not matter for this test data
             return 100;
         }
@@ -448,7 +448,7 @@ public final class ArrowTestUtils {
         }
 
         @Override
-        public int sizeOf() {
+        public long sizeOf() {
             // Does not matter for this test data
             return 100;
         }
