+++ org.knime.core.columnar.arrow/src/main/java/org/knime/core/columnar/arrow/data/ArrowDoubleData.java
@@ -93,7 +93,7 @@ public final class ArrowDoubleData {
         }
 
         @Override
-        public int sizeOf() {
+        public long sizeOf() {
             return ArrowSizeUtils.sizeOfFixedWidth(m_vector);
         }
 
@@ -127,7 +127,7 @@ public final class ArrowDoubleData {
         }
 
         @Override
-        public int sizeOf() {
+        public long sizeOf() {
             return ArrowSizeUtils.sizeOfFixedWidth(m_vector);
         }
     }
