+++ org.knime.core.columnar.arrow/src/main/java/org/knime/core/columnar/arrow/data/ArrowFloatData.java
@@ -93,7 +93,7 @@ public final class ArrowFloatData {
         }
 
         @Override
-        public int sizeOf() {
+        public long sizeOf() {
             return ArrowSizeUtils.sizeOfFixedWidth(m_vector);
         }
 
@@ -126,7 +126,7 @@ public final class ArrowFloatData {
         }
 
         @Override
-        public int sizeOf() {
+        public long sizeOf() {
             return ArrowSizeUtils.sizeOfFixedWidth(m_vector);
         }
     }
